requirements.txt
gym 
neat-python